Update new design for BalAlert
