+++ src/pages/DexV2/Pool/AddLiquidity/AddLiquidityForm/components/AddLiquidityPreview/components/ActionSteps.tsx
@@ -10,7 +10,7 @@ import Loader from 'components/Loader'
 import useEthers from 'hooks/dex-v2/useEthers'
 import { TransactionAction, postConfirmationDelay } from 'hooks/dex-v2/useTransactions'
 import { dateTimeLabelFor } from 'hooks/dex-v2/useTime'
-import { BalAlert } from 'pages/DexV2/common/BalAlert'
+import BalAlert from 'pages/DexV2/common/BalAlert'
 import { useSwapState } from 'state/dexV2/swap/useSwapState'
 import { usePoolState } from 'state/dexV2/pool/usePoolState'
 import { setActionStates } from 'state/dexV2/pool'
