+++ src/pages/DexV2/common/BalAlert.tsx
@@ -1,204 +1,149 @@
-// BalAlert.tsx
-import React from 'react';
-import styled from 'styled-components';
-import { AlertCircle, Sun } from 'react-feather';
+import React, { useMemo } from 'react'
+import styled from 'styled-components'
 
-// Define the allowed types and sizes.
-export type AlertType = 'warning' | 'error' | 'info' | 'tip';
-export type SizeType = 'sm' | 'md' | 'lg';
+export type AlertType = 'warning' | 'error' | 'info' | 'tip'
 
-interface BalAlertProps {
-  className?: string;
-  type?: AlertType;
-  size?: SizeType;
-  title?: React.ReactNode;
-  description?: React.ReactNode;
-  actionLabel?: string;
-  raised?: boolean;
-  block?: boolean;
-  contentClass?: string;
-  square?: boolean;
-  noBorder?: boolean;
-  dismissable?: boolean;
-  onActionClick?: () => void;
-  onClose?: () => void;
-  children?: React.ReactNode; // used as default slot content
+export interface BalAlertProps extends React.HTMLAttributes<HTMLDivElement> {
+  type?: AlertType
+  size?: 'sm' | 'md' | 'lg'
+  title?: string
+  description?: string
+  actionLabel?: string
+  raised?: boolean
+  block?: boolean
+  contentClass?: string
+  square?: boolean
+  noBorder?: boolean
+  dismissable?: boolean
+  onActionClick?: () => void
+  onClose?: () => void
+  // Optional slot override for the title
+  titleSlot?: React.ReactNode
+  // Default slot for description / extra content
+  children?: React.ReactNode
 }
 
-// ----- Styled Components ----- //
+/* Styled Components */
+
+interface AlertWrapperProps {
+  bgColor: string
+  borderColor: string
+  padding: string
+  raised?: boolean
+  block?: boolean
+  rounded?: boolean
+  noBorder?: boolean
+}
+
+const AlertWrapper = styled.div<AlertWrapperProps>`
+  position: relative;
+  display: inline-block;
+  font-weight: 500;
+  background-color: ${(p) => p.bgColor};
+  padding: ${(p) => p.padding};
+  ${(p) => !p.noBorder && `border: 1px solid ${p.borderColor};`}
+  ${(p) =>
+    p.raised &&
+    `
+    box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+  `}
+  ${(p) => p.block && `width: 100%;`}
+  ${(p) => (p.rounded ? `border-radius: 0.5rem;` : '')}
+`
 
 interface AlertContainerProps {
-  alertType: AlertType;
-  size: SizeType;
-  raised: boolean;
-  block: boolean;
-  square: boolean;
-  noBorder: boolean;
+  alignItems: string
 }
 
 const AlertContainer = styled.div<AlertContainerProps>`
-  position: relative;
-  background-color: ${({ alertType }) => {
-    switch (alertType) {
-      case 'tip':
-        return '#eff6ff';
-      case 'warning':
-        return '#fffaf0';
-      case 'error':
-        return '#fff5f5';
-      default:
-        return '#f7fafc';
-    }
-  }};
-  border: ${({ noBorder, alertType }) =>
-    noBorder
-      ? 'none'
-      : `1px solid ${
-          alertType === 'tip'
-            ? '#bfdbfe'
-            : alertType === 'warning'
-            ? '#fed7aa'
-            : alertType === 'error'
-            ? '#feb2b2'
-            : '#edf2f7'
-        }`};
-  color: black;
-  padding: ${({ size }) => (size === 'sm' ? '4px' : '8px')};
-  box-shadow: ${({ raised }) =>
-    raised ? '0 1px 2px rgba(0,0,0,0.05)' : 'none'};
-  width: ${({ block }) => (block ? '100%' : 'auto')};
-  border-radius: ${({ square }) => (square ? '0' : '8px')};
-`;
-
-interface InnerContainerProps {
-  centerContent: boolean;
-}
-
-const InnerContainer = styled.div<InnerContainerProps>`
   display: flex;
   flex-grow: 1;
-  align-items: ${({ centerContent }) => (centerContent ? 'center' : 'flex-start')};
-`;
+  align-items: ${(p) => p.alignItems};
+`
 
-interface IconContainerProps {
-  size: SizeType;
-  alertType: AlertType;
+interface IconWrapperProps {
+  size: string
+  color: string
 }
 
-const getIconColors = (alertType: AlertType) => {
-  switch (alertType) {
-    case 'tip':
-      return { color: '#2b6cb0', backgroundColor: 'transparent' };
-    case 'warning':
-      return { color: '#dd6b20', backgroundColor: 'rgba(234, 88, 12, 0.1)' };
-    case 'error':
-      return { color: '#f56565', backgroundColor: 'rgba(245, 101, 101, 0.1)' };
-    default:
-      return { color: '#718096', backgroundColor: 'rgba(0, 0, 0, 0.1)' };
-  }
-};
-
-const IconContainer = styled.div<IconContainerProps>`
+const IconWrapper = styled.div<IconWrapperProps>`
+  border-radius: 50%;
   display: flex;
   align-items: center;
   justify-content: center;
-  margin-right: 8px;
-  border-radius: 9999px;
-  width: ${({ size }) =>
-    size === 'sm' ? '20px' : size === 'lg' ? '28px' : '24px'};
-  height: ${({ size }) =>
-    size === 'sm' ? '20px' : size === 'lg' ? '28px' : '24px'};
-  color: ${({ alertType }) => getIconColors(alertType).color};
-  background-color: ${({ alertType }) => getIconColors(alertType).backgroundColor};
-`;
+  margin-right: 0.5rem;
+  width: ${(p) => p.size};
+  height: ${(p) => p.size};
+  color: ${(p) => p.color};
+`
 
-interface ContentContainerProps {
-  centerContent: boolean;
-  flexCol: boolean;
+interface ContentWrapperProps {
+  flexDirection: string
 }
 
-const ContentContainer = styled.div<ContentContainerProps>`
+const ContentWrapper = styled.div<ContentWrapperProps>`
   display: flex;
   flex-grow: 1;
+  flex-direction: ${(p) => p.flexDirection};
   min-width: 0;
-  align-items: ${({ centerContent }) => (centerContent ? 'center' : 'flex-start')};
-  flex-direction: ${({ flexCol }) => (flexCol ? 'column' : 'row')};
-`;
+`
 
 interface TitleProps {
-  size: SizeType;
-  hasDescription: boolean;
+  fontSize: string
+  fontWeight: number
+  titleColor: string
 }
 
 const Title = styled.h5<TitleProps>`
   margin: 0;
-  font-size: ${({ size }) =>
-    size === 'sm' ? '0.75rem' : size === 'lg' ? '1rem' : '0.875rem'};
-  font-weight: ${({ hasDescription }) => (hasDescription ? 600 : 400)};
-`;
+  color: ${(p) => p.titleColor};
+  font-size: ${(p) => p.fontSize};
+  font-style: normal;
+  font-weight: ${(p) => p.fontWeight};
+  line-height: 150%;
+  letter-spacing: -0.28px;
+`
 
 interface DescriptionProps {
-  size: SizeType;
-  alertType: AlertType;
+  fontSize: string
+  color: string
 }
 
 const Description = styled.div<DescriptionProps>`
-  margin: 0;
-  font-size: ${({ size }) =>
-    size === 'sm' ? '0.75rem' : size === 'lg' ? '1rem' : '0.875rem'};
-  font-weight: 400;
-  color: ${({ alertType }) =>
-    alertType === 'tip' ? 'black' : 'rgba(0, 0, 0, 0.7)'};
-  white-space: pre-wrap;
   overflow: hidden;
   word-break: break-word;
-`;
+  white-space: pre-wrap;
+  font-weight: normal;
+  color: rgba(41, 41, 51, 0.5);
+  font-family: Inter;
+  font-size: ${(p) => p.fontSize};
+  font-style: normal;
+  line-height: 150%;
+  letter-spacing: -0.28px;
+`
 
-interface ActionContainerProps {
-  hasDescription: boolean;
-}
+const ActionWrapper = styled.div`
+  display: flex;
+`
 
-const ActionContainer = styled.div<ActionContainerProps>`
-  ${({ hasDescription }) =>
-    hasDescription ? 'margin-top: 4px;' : 'padding-left: 16px;'}
-`;
-
-interface ActionButtonProps {
-  btnColor: 'yellow' | 'red' | 'gray';
-}
-
-const ActionButton = styled.button<ActionButtonProps>`
-  padding: 4px 8px;
+const ActionButton = styled.button`
+  background-color: transparent;
+  border: none;
+  color: blue;
   font-size: 0.75rem;
-  border: none;
-  background-color: ${({ btnColor }) => {
-    switch (btnColor) {
-      case 'yellow':
-        return '#f6e05e';
-      case 'red':
-        return '#fc8181';
-      default:
-        return '#e2e8f0';
-    }
-  }};
-  color: black;
-  border-radius: 4px;
   cursor: pointer;
-`;
+`
 
-const DismissButton = styled.button`
+const DismissButtonWrapper = styled.div`
   position: absolute;
-  top: 4px;
-  right: 4px;
-  background: transparent;
-  border: none;
+  top: 0.25rem;
+  right: 0.25rem;
   cursor: pointer;
-  font-size: 1rem;
-`;
+`
 
-// ----- The BalAlert Component ----- //
+/* Main Component */
 
-export const BalAlert: React.FC<BalAlertProps> = ({
+const BalAlert: React.FC<BalAlertProps> = ({
   type = 'info',
   size = 'md',
   title = 'A title message',
@@ -212,72 +157,123 @@ export const BalAlert: React.FC<BalAlertProps> = ({
   dismissable = false,
   onActionClick,
   onClose,
-  className,
+  titleSlot,
   children,
 }) => {
-  // Determine if there is any extra content (description or children)
-  const hasContent = Boolean(description || children);
-  // Center the container when no extra content exists.
-  const centerContainer = !hasContent;
-  // Use a column layout if extra content exists.
-  const flexCol = hasContent;
+  // Compute color and spacing values (remove dark mode colors)
+  const bgColor = useMemo(() => {
+    switch (type) {
+      case 'tip':
+        return '#ebf8ff' // blue-50
+      case 'warning':
+        return 'rgba(255, 190, 24, 0.05)' // custom warning background
+      case 'error':
+        return '#fff5f5' // red-50
+      default:
+        return '#f7fafc' // gray-100
+    }
+  }, [type])
 
-  // Determine button color based on type.
-  let btnColor: 'yellow' | 'red' | 'gray';
-  if (type === 'warning') {
-    btnColor = 'yellow';
-  } else if (type === 'error') {
-    btnColor = 'red';
-  } else {
-    btnColor = 'gray';
-  }
+  const borderColor = useMemo(() => {
+    switch (type) {
+      case 'tip':
+        return '#bee3f8' // blue-200
+      case 'warning':
+        return 'rgba(255, 190, 24, 0.50)' // custom warning border
+      case 'error':
+        return '#feb2b2' // red-200
+      default:
+        return '#e2e8f0' // gray-200
+    }
+  }, [type])
 
-  // Compute icon size (matching the IconContainer dimensions).
-  const computedIconSize = size === 'sm' ? 20 : size === 'lg' ? 28 : 24;
+  const titleColor = useMemo(() => {
+    switch (type) {
+      case 'tip':
+        return '#2b6cb0' // blue-700
+      case 'warning':
+        return '#ffbe18' // orange-500
+      case 'error':
+        return '#FF6565' // red-500
+      default:
+        return '#4a5568' // secondary
+    }
+  }, [type])
+
+  const paddingVal = size === 'sm' ? '0.25rem' : '0.5rem'
+  const fontSize = size === 'sm' ? '0.75rem' : size === 'lg' ? '1rem' : '0.875rem'
+  const iconSizeVal = size === 'sm' ? '20px' : size === 'lg' ? '28px' : '24px'
+
+  const iconColorVal = useMemo(() => {
+    switch (type) {
+      case 'tip':
+        return '#2b6cb0' // blue-700
+      case 'warning':
+        return '#ed8936' // orange-500
+      case 'error':
+        return '#f56565' // red-500
+      default:
+        return '#4a5568' // secondary
+    }
+  }, [type])
+
+  // If a description or children is provided, we assume the content should stack vertically.
+  const flexDirection = description || children ? 'column' : 'row'
+  const alignItems = description || children ? 'flex-start' : 'center'
+  const titleFontWeight = description || children ? 600 : 400
+  const descColor = type === 'tip' ? '#000' : 'rgba(0, 0, 0, 0.7)'
 
   return (
-    <AlertContainer
-      alertType={type}
-      size={size}
+    <AlertWrapper
+      bgColor={bgColor}
+      borderColor={borderColor}
+      padding={paddingVal}
       raised={raised}
       block={block}
-      square={square}
+      rounded={!square}
       noBorder={noBorder}
-      className={className}
     >
-      <InnerContainer centerContent={centerContainer}>
-        <IconContainer size={size} alertType={type}>
+      <AlertContainer alignItems={alignItems}>
+        <IconWrapper size={iconSizeVal} color={iconColorVal}>
           {type === 'tip' ? (
-            <Sun size={computedIconSize} />
+            // Replace with your LightBulbIcon if available
+            <span style={{ fontSize: iconSizeVal }}>💡</span>
           ) : (
-            <AlertCircle size={computedIconSize} />
+            // Replace with your alert icon if available
+            <span style={{ fontSize: iconSizeVal }}>⚠️</span>
           )}
-        </IconContainer>
-        <ContentContainer centerContent={centerContainer} flexCol={flexCol} className={contentClass}>
+        </IconWrapper>
+        <ContentWrapper flexDirection={flexDirection} className={contentClass}>
           <div>
-            <Title size={size} hasDescription={hasContent}>
-              {title}
+            <Title fontSize={fontSize} fontWeight={titleFontWeight} titleColor={titleColor}>
+              {titleSlot || title}
             </Title>
-            {hasContent && (
-              <Description size={size} alertType={type}>
+            {(children || description) && (
+              <Description fontSize={fontSize} color={descColor}>
                 {children || description}
               </Description>
             )}
           </div>
           {actionLabel && (
-            <ActionContainer hasDescription={hasContent}>
-              <ActionButton btnColor={btnColor} onClick={onActionClick}>
-                {actionLabel}
-              </ActionButton>
-            </ActionContainer>
+            <ActionWrapper style={description || children ? { marginTop: '0.25rem' } : { paddingLeft: '1rem' }}>
+              <ActionButton onClick={onActionClick}>{actionLabel}</ActionButton>
+            </ActionWrapper>
           )}
-        </ContentContainer>
-      </InnerContainer>
+        </ContentWrapper>
+      </AlertContainer>
       {dismissable && (
-        <DismissButton onClick={onClose}>
-          ×
-        </DismissButton>
+        <DismissButtonWrapper
+          onClick={(e) => {
+            e.stopPropagation()
+            onClose && onClose()
+          }}
+        >
+          {/* Replace with your close icon if available */}
+          <span style={{ fontSize: '1rem' }}>×</span>
+        </DismissButtonWrapper>
       )}
-    </AlertContainer>
-  );
-};
+    </AlertWrapper>
+  )
+}
+
+export default BalAlert
